fix catalog not searching all_text_tsimv field
Now, when doing a "normal" catalog search, if a term is found in a child work's searchable_text field or a file set's all_text_tsimv field, the parent work will show up in the results.

This used to work, but seems to have been broken by recent slug work. Using slugs instead of IDs seems to fix the problem.

Co-authored-by: Alisha Evans <alishaevn2@gmail.com>
Co-authored-by: April Rieger <april.rieger@yahoo.com>
Co-authored-by: Shana Moore <shana@scientist.com>